requirements.txt
streamlit==1.41.1
langchain==0.3.14
langchain-openai==0.2.14
python-dotenv==1.2.2
